rearranged functions into df_utils.py and graph_utils.py
